Remove unboxing.

+++ testsrc/unit/leetcode/problems_0161_0170/Test_0169.java
@@ -8,8 +8,8 @@ class Test_0169 {
     @Test
     void itFindsTheMajorityElement() {
         var P = new Problem_0169();
-        assertEquals(3, P.majorityElement(new int[] { 3, 2, 3 }));
-        assertEquals(2, P.majorityElement(new int[] { 2, 2, 1, 1, 1, 2, 2 }));
-        assertEquals(3, P.majorityElement(new int[] { 3, 3, 4 }));
+        assertEquals(3, P.majorityElement(new int[] { 3, 2, 3 }).intValue());
+        assertEquals(2, P.majorityElement(new int[] { 2, 2, 1, 1, 1, 2, 2 }).intValue());
+        assertEquals(3, P.majorityElement(new int[] { 3, 3, 4 }).intValue());
     }
 }
